feat: 旅客商品

+++ src/main/java/com/yumefusaka/SmartAirport/pojo/VO/LuggageVO.java
@@ -10,7 +10,7 @@ public class LuggageVO {
 
     private String description;
 
-    private int passenger_id;
+    private long passenger_id;
 
     private int number;
